Add processing for attachments in SoapOutInterceptor
git-svn-id: https://svn.apache.org/repos/asf/incubator/cxf/trunk@436199 13f79535-47bb-0310-9956-ffa450edef68

+++ rt2/bindings/soap/src/main/java/org/objectweb/celtix/bindings/soap2/attachments/CachedOutputStream.java
@@ -13,13 +13,13 @@ import java.io.OutputStream;
 
 public class CachedOutputStream extends OutputStream {
     private OutputStream currentStream;
-    private int threshold;
+    private long threshold;
     private int totalLength;
     private boolean inmem;
     private File tempFile;
     private File outputDir;
 
-    public CachedOutputStream(int thresholdParam, File outputDirParam) throws IOException {
+    public CachedOutputStream(long thresholdParam, File outputDirParam) throws IOException {
         this.threshold = thresholdParam;
         this.outputDir = outputDirParam;
